server/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
colorama==0.4.4
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
django-bootstrap-pagination==1.7.1
django-bootstrap-v5==1.0.0
django-cors-headers==3.7.0
django-extensions==3.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.0.12
importlib-metadata==2.1.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.16
Pygments==2.8.1
PyJWT==1.7.1
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
traitlets==5.0.5
virtualenv==20.4.2
wcwidth==0.2.5
zipp==3.4.1
